-some more cleanup of things -added support for wiiflow internal boot method in neek2o, so you can use cheats, debugger etc if you boot channels in neek2o (doesnt work if you are in real nand and just select boot neek)

+++ trunk/resources/wiiflow_game_booter/source/disc.h
@@ -7,7 +7,7 @@ extern "C" {
 #endif /* __cplusplus */
 
 s32 Disc_Open();
-s32 Disc_FindPartition(u64 *outbuf);
+s32 Disc_FindPartition(u32 *outbuf);
 s32	Disc_SetUSB(const u8 *id, bool frag);
 void Disc_SetLowMem(u32 IOS);
 void Disc_SetTime();
